broke Messages Class up into bite sized chunks
